PCEUncertainty.jl

A package for Polynomial Chaos Expansions (PCE) in Uncertainty Quantification (UQ) Applications.

Function Documentation

PCEUncertainty.StochasticODEProblem — Type

StochasticODEProblem - structure to contain information about an ODE problem with random coefficient

Fields:

t_max - end time for the ODE simulation basis - data structure with the polynomial basis collocation_nodes - set of nodes to perform the collocation with collocation_weights - set of weights for the collocation nodes

PCEUncertainty.generate_collocation — Function

generate_collocation - compute the solution to the ODE for a set of collocation points in the probability space

Fields:

prob - data structure describing problem to be solved

PCEUncertainty.generate_collocation_2d — Function

generate_collocation_2d - compute the solution to the ODE for a set of collocation points in the probability space; different from generate_collocation because this encodes data about the underlying differential equation model.

Fields:

prob - data structure describing problem to be solved

PCEUncertainty.generate_pce_coefficients — Function

generate_pce_coefficients - compute the PCE coefficients for the orthogonal representation of the interpolant

Fields:

prob - data structure describing problem to be solved collocation_values - values of the solution of the ODE problem at the collocation nodes vandermonde - whether or not to compute the PCE coefficients by inverting a Vandermonde matrix

PCEUncertainty.evaluate_basis — Function

evaluate_basis - evaluate the polynomial basis (univariate or multivariate) at a vector of nodes and return a matrix with this information

Fields:

nodes - set of nodes to evaluate the basis at basis - vector of (univariate) basis functions. If length(basis) == 1, then the basis is univariate. Otherwise, the basis is a multivariate basis constructed as a tensor product of univariate bases.

PCEUncertainty.compute_pce_approximation — Function

compute_pce_approximation - compute the PCE approximation of the performance measure F at the quadrature nodes and the associated quadrature weights

Fields:

resolution - degree of the basis to be used in the interpolation of the performance measure vandermonde - whether or not to use inversion of the Vandermonde matrix for computation of PCE coefficients num_quadrature_nodes - number of nodes to use for computing the risk-sensitive forms

PCEUncertainty.compute_integrals — Function

compute_integrals - numerically evaluates the risk-sensitive form Λ and the two hybrid forms Λ¹ and Λ² for fixed values of c.

Fields:

quadrature_values - values of the performance measure at the quadrature nodes quadrature_weights - weights associated with the quadrature nodes c - numerical parameter associated with the integrals